reddit-fetcher/requirements.txt
praw==7.8.1
kafka-python==2.2.9